requirements.txt
kafka-python
pyzmq
pyopenssl
python-pkcs11
cryptography
